remove components/DropDown/types.ts
