+++ components/DropDown/DropDown.tsx
@@ -20,14 +20,19 @@
  */
 
 import { useTheme } from '@emotion/react';
-import { FC, useCallback, useEffect, useRef, useState } from 'react';
+import { FC, ReactChildren, ReactNode, useCallback, useEffect, useRef, useState } from 'react';
 import { DropdownButton } from '../Button';
 import { ChevronDown } from '../theme/icons';
+import DropDownContent from './DropDownContent';
 import DropDownItem from './DropDownItem';
 import { ChevronStyle, DropDownTitleStyle, DropdownMenuStyle, ParentStyle } from './styles';
-import { DropDownProps } from './types';
-import DropDownContent from './DropDownContent';
 
+type DropDownProps = {
+	title?: string;
+	leftIcon?: ReactNode;
+	MenuItemMap?: Map<string, string | (() => void)>;
+	children?: ReactNode | ReactChildren;
+};
 const Dropdown: FC<DropDownProps> = ({ MenuItemMap = new Map(), title, children, leftIcon }) => {
 	const [open, setOpen] = useState(false);
 	const dropdownRef = useRef<HTMLDivElement>(null);
