+++ components/DropDown/DropDownContent.tsx
@@ -19,8 +19,12 @@
  *
  */
 
-import { FC } from 'react';
-import { DropDownContentProps } from './types';
+import { FC, ReactChildren, ReactNode } from 'react';
+
+export type DropDownContentProps = {
+	label?: string;
+	children?: ReactNode | ReactChildren;
+};
 
 const DropDownContent: FC<DropDownContentProps> = ({ children, label }) => {
 	return <>{children || label}</>;
